Поменяли int на long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
 
     public static void main(String[] args) {
-        int[] sales = {20, 0, 190, 25, 191};
+        long[] sales = {20, 0, 190, 25, 191};
         SalesManager salesManager = new SalesManager(sales);
 
         System.out.println("Максимальный размер продаж: " + salesManager.max());
